@@ -2,20 +2,45 @@ name: kelas_belajar
 version: 1.0.0+1
 publish_to: none
 description: A new Flutter project.
-environment: 
-  sdk: '>=3.4.0 <4.0.0'
+environment:
+  sdk: ">=3.4.0 <4.0.0"
 
-dependencies: 
+dependencies:
+  connectivity_plus: ^5.0.2
   cupertino_icons: ^1.0.6
+  get_cli: ^1.8.1
   get: 4.6.6
-  flutter: 
+  flutter:
     sdk: flutter
+  flutter_svg: ^2.0.10+1
+  cached_network_image: ^3.3.1
+  path_provider: ^2.1.3
+  flutter_cache_manager: ^3.3.2
+  permission_handler: ^11.3.1
+  carousel_slider: ^4.2.1
+  font_awesome_flutter: ^10.7.0
+  flutter_easyloading: ^3.0.5
+  modal_bottom_sheet: ^3.0.0
+  file_picker: ^8.0.3
+  syncfusion_flutter_pdfviewer: ^24.2.8
+  image_picker: ^1.1.2
+  shimmer: ^3.0.0
+  google_fonts: ^6.2.1
+  uuid: ^4.4.0
+  package_info_plus: ^8.0.0
+  http: ^1.2.1
+  url_launcher: ^6.3.0
+  firebase_core: ^3.0.0
+  firebase_auth: ^5.0.0
+  firebase_messaging: ^15.0.0
+  google_sign_in: ^6.2.1
+  sign_in_with_apple: ^6.1.0
+  get_storage: ^2.1.1
 
-dev_dependencies: 
+dev_dependencies:
   flutter_lints: 4.0.0
-  flutter_test: 
+  flutter_test:
     sdk: flutter
 
-flutter: 
+flutter:
   uses-material-design: true
-
