@@ -2,22 +2,21 @@ import 'package:flutter/material.dart';
 
 import 'package:get/get.dart';
 
+import '../../infrastructure/widget/mahas_responsive.dart';
 import 'controllers/home.controller.dart';
+import 'screen/desktop.dart';
+import 'screen/mobile.dart';
+import 'screen/tablet.dart';
 
 class HomeScreen extends GetView<HomeController> {
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('HomeScreen'),
-        centerTitle: true,
-      ),
-      body: const Center(
-        child: Text(
-          'HomeScreen is working',
-          style: TextStyle(fontSize: 20),
-        ),
+    return const Scaffold(
+      body: MahasResponsive(
+        mobile: MobileLayout(),
+        tablet: TabletLayout(),
+        desktop: DesktopLayout(),
       ),
     );
   }
