@@ -1,8 +1,11 @@
 class Routes {
   static Future<String> get initialRoute async {
     // TODO: implement method
-    return HOME;
+    return BASE;
   }
 
+  static const BASE = '/base';
   static const HOME = '/home';
+  static const PROFILE = '/profile';
+  static const MODULE = '/module';
 }
